ihuayu_backend/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
packaging==23.1
python-decouple==3.8
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.3
SQLAlchemy==2.0.20
typing_extensions==4.8.0
Werkzeug==2.3.7
